+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2021/09/08
+  * 增加image\_manager\_set\_max\_mem\_size\_of\_cached\_images用于限制最大图片缓存占用的内存。
+
 2021/09/07
   * 完善fscript，没有定义的变量，不再当作字符串。 
   * 完善utf8（感谢智明提供补丁）